Revert "added the commands to the controllers"
This reverts commit 7f9fe51534ceac2031846f421dcfc690a3dab5b3.

+++ FoodOrderingApp-service/src/main/java/com/upgrad/FoodOrderingApp/service/entity/OrdersEntity.java
@@ -104,7 +104,7 @@ public class OrdersEntity implements Serializable {
         return bill;
     }
 
-    public void setBill(Double bill) {
+    public void setBill(Float bill) {
         this.bill = bill;
     }
